Replaced the shared signature cache by an instance-local one
Because the main purpose for this class is to act as a stand-in for a delegate, and delegates should conform to protocols, we can lookup the method signatures in the protocol.

This also solves the edge case where a protocol contains a method of the same name as some other type, but with a different signature, which might otherwise lead to data corruption.

LIMITATIONS:
Currently, the helper function only looks up the optional instance methods in the protocol.
This can be extended pretty easily.